refine agent directories. (#1109)
