+++ .github/workflows/docker/compose/agent-compose.yaml
@@ -3,7 +3,7 @@
 
 # this file should be run in the root of the repo
 services:
-  agent-langchain:
+  agent:
     build:
-      dockerfile: comps/agent/langchain/Dockerfile
-    image: ${REGISTRY:-opea}/agent-langchain:${TAG:-latest}
+      dockerfile: comps/agent/src/Dockerfile
+    image: ${REGISTRY:-opea}/agent:${TAG:-latest}
