+++ comps/agent/src/Dockerfile
@@ -21,19 +21,19 @@ COPY comps /home/user/comps
 
 RUN pip install --no-cache-dir --upgrade pip setuptools && \
     if [ ${ARCH} = "cpu" ]; then \
-        pip install --no-cache-dir --extra-index-url https://download.pytorch.org/whl/cpu -r /home/user/comps/agent/langchain/requirements.txt; \
+        pip install --no-cache-dir --extra-index-url https://download.pytorch.org/whl/cpu -r /home/user/comps/agent/src/requirements.txt; \
     else \
-        pip install --no-cache-dir -r /home/user/comps/agent/langchain/requirements.txt; \
+        pip install --no-cache-dir -r /home/user/comps/agent/src/requirements.txt; \
     fi
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
 
 USER root
 
-RUN mkdir -p /home/user/comps/agent/langchain/status && chown -R user /home/user/comps/agent/langchain/status
+RUN mkdir -p /home/user/comps/agent/src/status && chown -R user /home/user/comps/agent/src/status
 
 USER user
 
-WORKDIR /home/user/comps/agent/langchain/
+WORKDIR /home/user/comps/agent/src/
 
 ENTRYPOINT ["python", "agent.py"]
