+++ comps/agent/src/integrations/global_var.py
