+++ comps/agent/src/integrations/strategy/react/utils.py
