+++ comps/agent/src/integrations/strategy/sqlagent/hint.py
