+++ comps/agent/src/sql_agent.png
