+++ comps/agent/src/sql_agent_llama.png
